// run-archive/scenario-2/1777364720430-scenario-2-deepseek-v4-pro-plan-flash-build-mimo-review-attempt-1/frontend/src/generated/booking-hooks.ts
export interface Booking {
  id: string;
  restaurantId: string;
  restaurantName: string;
  tableId: string;
  partySize: number;
  date: string;
  time: string;
  guestName: string;
  guestEmail: string;
  /** @nullable */
  userId?: string | null;
}